設定ソースコードを分離/CeVIO GUI準備/冒とくフィルタのパラメータがバインドされていなかった問題を修正/--out_obs_text_starts_withオプション画面実装
